+++ CHANGES.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## Development version
+
+### Breaking changes
+### Bugfixes
+### New features
+
+ - Support nested GitLab projects, e.g. http://gitlab.com/FOO/BAR/BAZ.git.
+   Contribution from @sebastiaanspeck.
+
 ## Version 0.1.0
 
 Initial release.